NMS: Rebuild nmsInfoBox
Basically the old model forced each window to handle all the logic needed for
showing and managing a window. The new model takes care of the basic window-
management and lets us focus on "panels" that are reusable, independent of each
other, and can re-render themselves without concern for any other objects. We
are also able to define windows and views that use existing panels just by
providing a basic template to the windowHandler.

Does still have som bugs and missing functionality that needs to be addressed.